requirements.txt
asgiref==3.8.1
Brotli==1.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
cloudinary==1.36.0
crispy-tailwind==1.0.3
dj-database-url==2.3.0
Django==5.1.6
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-crispy-forms==2.3
django-debug-toolbar==5.0.1
Faker==33.1.0
idna==3.10
pillow==11.1.0
psycopg-binary==3.2.3
psycopg2==2.9.10
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
whitenoise==6.9.0
